WIP: Upgrade to Mingw-w64 9.0.0
In this release _WIN32_WINNT is changed from 0x502 to 0xa00. This wreaks
havoc on a lot of existing software, including both GDB and busybox-w32.

GDB: The Autoconf check for bcrypt.h in the embedded Gnulib is broken
and probably never worked in the first place, so Gnulib never tries to
use this header. When the check fails, it then checks if it's safe to
link against bcrypt.dll based on the value of _WIN32_WINNT. Before
Mingw-w64 9.0.0, this check also failed. So it used a run-time dynamic
symbol lookup fallback.

However, if the version check succeeds, which it does with 9.0.0, Gnulib
incorrectly links bcrypt.dll, resulting in a build failure. It would be
better to fix the original Autoconf input, but unfortunately this is
distributed already pre-compiled into a configure script. Regenerating
the configure script introduces other problems, so it's not worth it.

Instead, as a workaround pretend the second check failed and use the
fallback as before.

In typical GNU fashion, Gnulib uses 187 lines of code to do something
poorly where 7 lines of code would do it well. Instead of hacking around
BCryptGenRandom, Gnulib should just call RtlGenRandom. This could be a
whole lot simpler and less fragile.

busybox-w32: The higher _WIN32_WINNT value causes new declarations to
become visible which conflict with definitions in busybox-w32, mostly in
networking headers like winsock2.h. This will take a lot of work to fix,
and those fixes should be done upstream instead.

As a result, w64devkit will not use Mingw-w64 until these upstream
issues have been sorted. This branch is a work-in-progress placeholder.